frontend/requirements.txt
streamlit==1.45.1
requests==2.32.3
pillow==11.2.1